Query fedora for start and end dates for pid purge rather than the yaml file.

+++ delete_datastreams.py
@@ -34,14 +34,23 @@ def purge_a_dsid(pids):
     purge_url = credentials['fedoraurl'] + ":8080/fedora/objects/"
     deleted_pids = []
     for pid in pids:
-        url = purge_url + pid + '/datastreams/{0}/?&startDT={1}&endDT={2}&logMessage'.format(credentials['datastream'], credentials['start'], credentials['end'])
-        x = requests.delete(url, auth=(credentials['username'], credentials['password']))
-        if x.status_code == 200:
-            if len(repr(x.text)) > 4:
-                deleted_pids.append("{0}:  {1}".format(pid, repr(x.text)))
-        else:
-            print("{0} had a status code of: {1}".format(pid, x.status_code))
+        lastVersionDate = getLastVersionDate(pid)
+        if lastVersionDate:
+            number_of_versions = len(lastVersionDate) - 1
+            end = lastVersionDate[1]
+            if len(lastVersionDate) == 2:
+                start = lastVersionDate[1]
+            else:
+                start = lastVersionDate[number_of_versions]
+            url = '{0}{1}/datastreams/{2}/?&startDT={3}&endDT={4}&logMessage'.format(purge_url, pid, credentials['datastream'], start, end)
+            x = requests.delete(url, auth=(credentials['username'], credentials['password']))
+            if x.status_code == 200:
+                if len(repr(x.text)) > 4:
+                    deleted_pids.append("{0}:  {1}".format(pid, repr(x.text)))
+            else:
+                print("{0} had a status code of: {1}".format(pid, x.status_code))
     return deleted_pids
+
 
 def build_markdown_file(array_of_pids):
     markdown_file = open('deleted_pids.md', 'w')
@@ -54,8 +63,21 @@ def build_markdown_file(array_of_pids):
           "See deleted_pids.md for more details.\n".format(len(array_of_pids)))
 
 
+def getLastVersionDate(object):
+    url = '{0}:8080/fedora/objects/{1}/datastreams/{2}/history?format=xml'.format(credentials['fedoraurl'], object, credentials['datastream'])
+    s = requests.get(url, auth=(credentials['username'], credentials['password']))
+    if s.status_code == 200:
+        x = s.text
+        json_string = json.dumps(xmltodict.parse(x))
+        data = json.loads(json_string)
+        versions = []
+        for objects in data['datastreamHistory']['datastreamProfile']:
+            if type(objects) == dict:
+                versions.append(objects['dsCreateDate'])
+        return versions
+
+
 if __name__ == "__main__":
     pids_in_collection = get_collection_objects(find_objects_string, initial_token, collection_objects)
     deleted_objects = purge_a_dsid(pids_in_collection)
     build_markdown_file(deleted_objects)
-
